Re-prove match pattern optimization
Included the proofs of optimization rules applied to match patterns to
the more general form of the datatype.
Encountering an issue where termination cannot be proved as described
below.

It's unclear how to say for a generic type classes fixes such as
```isabelle
class T =
	fixes map :: "('a => 'a) => 'a => 'a"
```
How you would express a property about this that allows you to show
termination of mutually recursive calls, e.g.
```isabelle
fun f where "'a::T => 'a"
	"f = map (f)"
```
Since every parameter passed to the mapping function is smaller, e.g.
(map (\lambda x. x) e), x will always be smaller than e,
we know that it will terminate but it is hard to express
this property in Isabelle.