Swap variable generation order; add inline comments
This swaps the order of variable generation so that the outer most
variable is generated first (as opposed to innermost).
Meta-optimization relies on this implicit property to be effective
when optimizing out shared conditions.

Additionally add inline comments to the Rules data structure
and subsequent generated Java so that the rule source
is clear in the generated Java.